+++ src/core/markdown/adapters/renderable-bundle.ts
@@ -13,7 +13,7 @@ import { apply } from '#utils/fp';
 import { generateLink } from './generate-link';
 import { getTypeGroup } from '../../shared/utils';
 import { Type } from '@cparra/apex-reflection';
-import { ObjectMetadata } from '../../reflection/reflect-object-source';
+import { ObjectMetadata } from '../../reflection/sobject/reflect-sobject-source';
 
 export function parsedFilesToRenderableBundle(
   config: MarkdownGeneratorConfig,
